Rename helper module and consequent import changes.
